inits imageinfo.py, copies over read_image_info from utils
