+++ lablib/operators/base.py
@@ -1,0 +1,21 @@
+import logging
+from typing import Any
+from pathlib import Path
+
+
+class BaseOperator:
+    def __init__(self, *args, **kwargs):
+        self.log = logging.getLogger(__name__)
+        self.log.setLevel(logging.DEBUG)
+
+    def __getitem__(self, k: str) -> Any:
+        return getattr(self, k)
+
+    def __setitem__(self, k: str, v: Any) -> None:
+        if hasattr(self, k):
+            setattr(self, k, v)
+        else:
+            self.log.error(f"Cannot set {k}={v}. Key {k} not found.", stack_info=True)
+
+    def update_from_path(self, path: Path) -> None:
+        raise NotImplementedError("update_from_path method not implemented.")
