+++ lablib/utils.py
@@ -1,15 +1,16 @@
 from __future__ import annotations
 import os
-import subprocess
 import math
 import uuid
-from copy import deepcopy
+import logging
 from pathlib import Path
 from typing import List
 
 import opentimelineio as otio
 
-from .operators import ImageInfo
+
+log = logging.getLogger(__name__)
+log.setLevel(logging.DEBUG)
 
 
 class format_dict(dict):
@@ -17,157 +18,6 @@ class format_dict(dict):
 
     def __missing__(self, key) -> str:
         return self._placeholder
-
-
-def read_image_info(
-    path: str,
-    default_timecode: str = None,
-    default_fps: float = None,
-    default_par: float = None,
-    default_channels: int = None,
-) -> ImageInfo:
-    if default_timecode is None:
-        default_timecode = "01:00:00:01"
-    if default_fps is None:
-        default_fps = 24.0
-    if default_par is None:
-        default_par = 1.0
-    if default_channels is None:
-        default_channels = 3
-
-    abspath = Path(path).as_posix()
-
-    result = {
-        "filename": abspath,
-        "origin_x": None,
-        "origin_y": None,
-        "width": None,
-        "height": None,
-        "display_width": None,
-        "display_height": None,
-        "channels": None,
-        "fps": None,
-        "par": None,
-        "timecode": None,
-    }
-
-    iinfo_res = deepcopy(result)
-    ffprobe_res = deepcopy(result)
-
-    iinfo_cmd = ["iinfo", "-v", abspath]
-
-    ffprobe_cmd = [
-        "ffprobe",
-        "-v",
-        "error",
-        "-select_streams",
-        "v:0",
-        "-show_entries",
-        "format_tags=timecode:stream_tags=timecode:stream=width,height,r_frame_rate,sample_aspect_ratio",
-        "-of",
-        "default=noprint_wrappers=1",
-        abspath,
-    ]
-
-    iinfo_out = (
-        subprocess.run(iinfo_cmd, capture_output=True, text=True)
-        .stdout.strip()
-        .splitlines()
-    )
-
-    ffprobe_out = (
-        subprocess.run(ffprobe_cmd, capture_output=True, text=True)
-        .stdout.strip()
-        .splitlines()
-    )
-
-    for l in iinfo_out:
-        if abspath in l and l.find(abspath) < 2:
-            vars = l.split(": ")[1].split(",")
-            size = vars[0].strip().split("x")
-            channels = vars[1].strip().split(" ")
-            iinfo_res.update(
-                {
-                    "width": int(size[0].strip()),
-                    "height": int(size[1].strip()),
-                    "display_width": int(size[0].strip()),
-                    "display_height": int(size[1].strip()),
-                    "channels": int(channels[0].strip()),
-                }
-            )
-        if "FramesPerSecond" in l or "framesPerSecond" in l:
-            vars = l.split(": ")[1].strip().split(" ")[0].split("/")
-            iinfo_res.update(
-                {"fps": float(round(float(int(vars[0]) / int(vars[1])), 3))}
-            )
-        if "full/display size" in l:
-            size = l.split(": ")[1].split("x")
-            iinfo_res.update(
-                {
-                    "display_width": int(size[0].strip()),
-                    "display_height": int(size[1].strip()),
-                }
-            )
-        if "pixel data origin" in l:
-            origin = l.split(": ")[1].strip().split(",")
-            iinfo_res.update(
-                {
-                    "origin_x": int(origin[0].replace("x=", "").strip()),
-                    "origin_y": int(origin[1].replace("y=", "").strip()),
-                }
-            )
-        if "smpte:TimeCode" in l:
-            iinfo_res["timecode"] = l.split(": ")[1].strip()
-        if "PixelAspectRatio" in l:
-            iinfo_res["par"] = float(l.split(": ")[1].strip())
-
-    for l in ffprobe_out:
-        vars = l.split("=")
-        if "width" in vars[0]:
-            ffprobe_res["display_width"] = int(vars[1].strip())
-        if "height" in vars[0]:
-            ffprobe_res["display_height"] = int(vars[1].strip())
-        if "r_frame_rate" in vars[0]:
-            rate = vars[1].split("/")
-            ffprobe_res["fps"] = float(
-                round(float(int(rate[0].strip()) / int(rate[1].strip())), 3)
-            )
-        if "timecode" in l:
-            ffprobe_res["timecode"] = vars[1]
-        if "sample_aspect_ratio" in l:
-            par = vars[1].split(":")
-            if vars[1] != "N/A":
-                ffprobe_res["par"] = float(int(par[0].strip()) / int(par[1].strip()))
-            else:
-                ffprobe_res["par"] = 1
-
-    for k, v in iinfo_res.items():
-        ffprobe_value = ffprobe_res[k]
-        if v == ffprobe_value:
-            result[k] = v
-        elif v and not ffprobe_value:
-            result[k] = v
-        elif ffprobe_value and not v:
-            result[k] = ffprobe_value
-        else:
-            result[k] = None
-
-    for key, default_value in (
-        ("width", result["display_width"]),
-        ("height", result["display_height"]),
-        ("origin_x", 0),
-        ("origin_y", 0),
-        ("timecode", default_timecode),
-        ("fps", default_fps),
-        ("par", default_par),
-        ("channels", default_channels),
-    ):
-        if not result[key]:
-            result[key] = default_value
-
-    image_info = ImageInfo(**result)
-
-    return image_info
 
 
 def offset_timecode(tc: str, frame_offset: int = None, fps: float = None) -> str:
